Move dataset files into pooling dir and update pathing
