+++ .gitignore
@@ -178,4 +178,4 @@ cython_debug/
 *.jpg
 *.png
 experiments/*
-data/
+datasets/
